+++ minijinja/src/filters.rs
@@ -107,6 +107,7 @@ pub(crate) fn get_builtin_filters() -> BTreeMap<&'static str, BoxedFilter> {
         rv.insert("length", BoxedFilter::new(length));
         rv.insert("count", BoxedFilter::new(length));
         rv.insert("dictsort", BoxedFilter::new(dictsort));
+        rv.insert("items", BoxedFilter::new(items));
         rv.insert("reverse", BoxedFilter::new(reverse));
         rv.insert("trim", BoxedFilter::new(trim));
         rv.insert("join", BoxedFilter::new(join));
@@ -212,6 +213,8 @@ mod builtins {
     }
 
     /// Dict sorting functionality.
+    ///
+    /// This filter works like `|items` but sorts the pairs by key first.
     #[cfg_attr(docsrs, doc(cfg(feature = "builtins")))]
     pub fn dictsort(_state: &State, v: Value) -> Result<Value, Error> {
         let mut pairs = match v.0 {
@@ -229,6 +232,41 @@ mod builtins {
                 .into_iter()
                 .map(|(k, v)| vec![Value::from(k.clone()), v.clone()])
                 .collect::<Vec<_>>(),
+        ))
+    }
+
+    /// Returns a list of pairs (items) from a mapping.
+    ///
+    /// This can be used to iterate over keys and values of a mapping
+    /// at once.  Note that this will use the original order of the map
+    /// which is typically arbitrary unless the `preserve_order` feature
+    /// is used in which case the original order of the map is retained.
+    /// It's generally better to use `|dictsort` which sorts the map by
+    /// key before iterating.
+    ///
+    /// ```text,ignore
+    /// <dl>
+    /// {% for key, value in my_dict|items %}
+    ///   <dt>{{ key }}
+    ///   <dd>{{ value }}
+    /// {% endfor %}
+    /// </dl>
+    /// ```
+    #[cfg_attr(docsrs, doc(cfg(feature = "builtins")))]
+    pub fn items(_state: &State, v: Value) -> Result<Value, Error> {
+        Ok(Value::from(
+            match v.0 {
+                ValueRepr::Map(ref v) => v.iter().collect::<Vec<_>>(),
+                _ => {
+                    return Err(Error::new(
+                        ErrorKind::ImpossibleOperation,
+                        "cannot convert value into pair list",
+                    ))
+                }
+            }
+            .into_iter()
+            .map(|(k, v)| vec![Value::from(k.clone()), v.clone()])
+            .collect::<Vec<_>>(),
         ))
     }